Move returns_json to senaite.core.decorators with a BBB shim
Consolidates returns_json alongside the other JSON endpoint decorators in
senaite.core.decorators. bika.lims.decorators re-exports it so existing
imports (from bika.lims.decorators import returns_json) keep working.
