+++ src/senaite/core/browser/label/api.py
@@ -19,11 +19,11 @@
 # Some rights reserved, see README and LICENSE.
 
 from bika.lims import api
-from bika.lims.decorators import returns_json
 from senaite.core.api import label as label_api
 from senaite.core.browser.views import JSONView
 from senaite.core.config.labels import SAMPLE_LABEL_REINDEX
 from senaite.core.decorators import json_require_permission
+from senaite.core.decorators import returns_json
 from senaite.core.permissions import ManageLabels
